Make it easier to subclass a remote ActiveSession
