+++ java/server/src/org/openqa/selenium/remote/server/ServicedSession.java
@@ -19,44 +19,19 @@ package org.openqa.selenium.remote.server;
 
 
 import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.openqa.selenium.remote.Dialect.OSS;
-
-import com.google.common.base.Preconditions;
-import com.google.common.base.StandardSystemProperty;
-import com.google.common.collect.ImmutableMap;
 
 import org.openqa.selenium.Capabilities;
-import org.openqa.selenium.ImmutableCapabilities;
 import org.openqa.selenium.SessionNotCreatedException;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.io.TemporaryFilesystem;
 import org.openqa.selenium.net.PortProber;
-import org.openqa.selenium.remote.Augmenter;
-import org.openqa.selenium.remote.Command;
-import org.openqa.selenium.remote.CommandCodec;
-import org.openqa.selenium.remote.CommandExecutor;
 import org.openqa.selenium.remote.Dialect;
-import org.openqa.selenium.remote.DriverCommand;
-import org.openqa.selenium.remote.ProtocolHandshake;
-import org.openqa.selenium.remote.RemoteWebDriver;
-import org.openqa.selenium.remote.Response;
-import org.openqa.selenium.remote.ResponseCodec;
 import org.openqa.selenium.remote.SessionId;
-import org.openqa.selenium.remote.http.HttpClient;
 import org.openqa.selenium.remote.http.HttpMethod;
 import org.openqa.selenium.remote.http.HttpRequest;
 import org.openqa.selenium.remote.http.HttpResponse;
-import org.openqa.selenium.remote.http.JsonHttpCommandCodec;
-import org.openqa.selenium.remote.http.JsonHttpResponseCodec;
-import org.openqa.selenium.remote.http.W3CHttpCommandCodec;
-import org.openqa.selenium.remote.http.W3CHttpResponseCodec;
-import org.openqa.selenium.remote.internal.ApacheHttpClient;
 import org.openqa.selenium.remote.server.jmx.JMXHelper;
-import org.openqa.selenium.remote.server.jmx.ManagedAttribute;
 import org.openqa.selenium.remote.server.jmx.ManagedService;
 import org.openqa.selenium.remote.service.DriverService;
 
-import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.Method;
 import java.net.URL;
@@ -65,24 +40,14 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.logging.Level;
-import java.util.logging.Logger;
 
 import javax.management.MalformedObjectNameException;
 import javax.management.ObjectName;
 
 @ManagedService
-public class ServicedSession implements ActiveSession {
-
-  private static Logger log = Logger.getLogger(ActiveSession.class.getName());
+public class ServicedSession extends RemoteSession {
 
   private final DriverService service;
-  private final SessionId id;
-  private final Dialect downstream;
-  private final Dialect upstream;
-  private final SessionCodec codec;
-  private final Map<String, Object> capabilities;
-  private final TemporaryFilesystem filesystem;
-  private final WebDriver driver;
 
   public ServicedSession(
       DriverService service,
@@ -91,63 +56,11 @@ public class ServicedSession implements ActiveSession {
       SessionCodec codec,
       SessionId id,
       Map<String, Object> capabilities) {
+    super(downstream, upstream, codec, id, capabilities);
+
     this.service = service;
-    this.downstream = downstream;
-    this.upstream = upstream;
-    this.codec = codec;
-    this.id = id;
-    this.capabilities = capabilities;
-
-    File tempRoot = new File(StandardSystemProperty.JAVA_IO_TMPDIR.value(), id.toString());
-    Preconditions.checkState(tempRoot.mkdirs());
-    this.filesystem = TemporaryFilesystem.getTmpFsBasedOn(tempRoot);
-
-    CommandExecutor executor = new ActiveSessionCommandExecutor(this);
-    this.driver = new Augmenter().augment(new RemoteWebDriver(
-        executor,
-        new ImmutableCapabilities(getCapabilities())));
 
     new JMXHelper().register(this);
-  }
-
-  @Override
-  public void execute(HttpRequest req, HttpResponse resp) throws IOException {
-    codec.handle(req, resp);
-  }
-
-  @Override
-  @ManagedAttribute
-  public SessionId getId() {
-    return id;
-  }
-
-  @Override
-  @ManagedAttribute
-  public Dialect getUpstreamDialect() {
-    return upstream;
-  }
-
-  @Override
-  @ManagedAttribute
-  public Dialect getDownstreamDialect() {
-    return downstream;
-  }
-
-  @Override
-  @ManagedAttribute
-  public Map<String, Object> getCapabilities() {
-    return capabilities;
-  }
-
-  @Override
-  @ManagedAttribute
-  public WebDriver getWrappedDriver() {
-    return driver;
-  }
-
-  @Override
-  public TemporaryFilesystem getFileSystem() {
-    return filesystem;
   }
 
   @Override
@@ -164,7 +77,7 @@ public class ServicedSession implements ActiveSession {
     service.stop();
   }
 
-  public static class Factory implements SessionFactory {
+  public static class Factory extends RemoteSession.Factory<DriverService> {
 
     private final Function<Capabilities, ? extends DriverService> createService;
     private final String serviceClassName;
@@ -176,9 +89,9 @@ public class ServicedSession implements ActiveSession {
             Class.forName(serviceClassName).asSubclass(DriverService.class);
 
         Function<Capabilities, ? extends DriverService> factory =
-            get(driverClazz, "createDefaultService", Capabilities.class);
+            get(driverClazz, Capabilities.class);
         if (factory == null) {
-          factory = get(driverClazz, "createDefaultService");
+          factory = get(driverClazz);
         }
 
         if (factory == null) {
@@ -195,10 +108,9 @@ public class ServicedSession implements ActiveSession {
 
     private Function<Capabilities, ? extends DriverService> get(
         Class<? extends DriverService> driverServiceClazz,
-        String methodName,
         Class... args) {
       try {
-        Method serviceMethod = driverServiceClazz.getDeclaredMethod(methodName, args);
+        Method serviceMethod = driverServiceClazz.getDeclaredMethod("createDefaultService", args);
         serviceMethod.setAccessible(true);
         return caps -> {
           try {
@@ -218,7 +130,9 @@ public class ServicedSession implements ActiveSession {
     }
 
     @Override
-    public Optional<ActiveSession> apply(Set<Dialect> downstreamDialects, Capabilities capabilities) {
+    public Optional<ActiveSession> apply(
+        Set<Dialect> downstreamDialects,
+        Capabilities capabilities) {
       DriverService service = createService.apply(capabilities);
 
       try {
@@ -228,74 +142,34 @@ public class ServicedSession implements ActiveSession {
 
         URL url = service.getUrl();
 
-        HttpClient client = new ApacheHttpClient.Factory().createClient(url);
-
-        Command command = new Command(null, DriverCommand.NEW_SESSION,
-                                      ImmutableMap.of("desiredCapabilities", capabilities));
-
-        ProtocolHandshake.Result result = new ProtocolHandshake().createSession(client, command);
-
-        SessionCodec codec;
-        Dialect upstream = result.getDialect();
-        Dialect downstream;
-        if (downstreamDialects.contains(result.getDialect())) {
-          codec = new Passthrough(url);
-          downstream = upstream;
-        } else {
-          downstream = downstreamDialects.isEmpty() ? OSS : downstreamDialects.iterator().next();
-
-          codec = new ProtocolConverter(
-              url,
-              getCommandCodec(downstream),
-              getResponseCodec(downstream),
-              getCommandCodec(upstream),
-              getResponseCodec(upstream));
-        }
-
-        Response response = result.createResponse();
-        //noinspection unchecked
-        return Optional.of(new ServicedSession(
-            service,
-            downstream,
-            upstream,
-            codec,
-            new SessionId(response.getSessionId()),
-            (Map<String, Object>) response.getValue()));
+        return performHandshake(service, url, downstreamDialects, capabilities);
       } catch (IOException | IllegalStateException | NullPointerException e) {
         log.log(Level.WARNING, e.getMessage(), e);
-        throw new SessionNotCreatedException("Cannot establish new session: " + e.getMessage(), e);
+        service.stop();
+        return Optional.empty();
       }
+    }
+
+    @Override
+    protected ServicedSession newActiveSession(
+        DriverService service,
+        Dialect downstream,
+        Dialect upstream,
+        SessionCodec codec,
+        SessionId id,
+        Map<String, Object> capabilities) {
+      return new ServicedSession(
+          service,
+          downstream,
+          upstream,
+          codec,
+          id,
+          capabilities);
     }
 
     @Override
     public String toString() {
       return getClass() + " (provider: " + serviceClassName + ")";
-    }
-
-    private CommandCodec<HttpRequest> getCommandCodec(Dialect dialect) {
-      switch (dialect) {
-        case OSS:
-          return new JsonHttpCommandCodec();
-
-        case W3C:
-          return new W3CHttpCommandCodec();
-
-        default:
-          throw new IllegalStateException("Unknown dialect: " + dialect);
-      }
-    }
-
-    private ResponseCodec<HttpResponse> getResponseCodec(Dialect dialect) {
-      switch (dialect) {
-        case OSS:
-          return new JsonHttpResponseCodec();
-
-        case W3C:
-          return new W3CHttpResponseCodec();
-
-        default:
-          throw new IllegalStateException("Unknown dialect: " + dialect);
-      }
     }
   }
